requirements.txt
sentence-transformers
scikit-learn
langchain
langchain-community
langchain-pinecone
pinecone-client
openai
python-dotenv
PyGithub
GitPython
